+++ typescript/ui/src/main.tsx
@@ -15,6 +15,7 @@ import { ChatHistoryPlugin } from "./plugins/ChatHistoryPlugin/index";
 import { BrowserRouter } from "react-router";
 import { Routes } from "./core/components/Routes.tsx";
 import { AuthPlugin } from "./plugins/AuthPlugin/index.tsx";
+import { UsagePlugin } from "./plugins/UsagePlugin/index.tsx";
 
 //Register plugins
 pluginManager.register(FeedbackFormPlugin);
@@ -22,6 +23,7 @@ pluginManager.register(ChatOptionsPlugin);
 pluginManager.register(SharePlugin);
 pluginManager.register(ChatHistoryPlugin);
 pluginManager.register(AuthPlugin);
+pluginManager.register(UsagePlugin);
 
 // Preload icons
 loadIcons([